refactor(co2): use .container filename suffix instead of .sfc
